+++ Sources/SwiftUIKit/Presentation/View+FullScreenCover.swift
@@ -8,7 +8,7 @@
 
 import SwiftUI
 
-@available(iOS 14.0, *)
+@available(iOS 14.0, tvOS 14.0, watchOS 7.0, *)
 public extension View {
     
     /**